third_party/blink/web_tests/virtual/mutation-events-enabled/dom/mutation-event-tests/html/details_summary/mutation-details-focus-expected.txt

CONSOLE WARNING: Listener added for a 'DOMNodeRemovedFromDocument' mutation event. This event type is no longer supported, and will be removed from this browser VERY soon. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
PASS unless crash
PASS successfullyParsed is true

TEST COMPLETE
